Absorb rules from defunct eslint-config-react-app
Our extension of the eslint-config-react-app has always caused problems:

1. Conflicts between versions of plugins
2. Unnecessary ESLint plugins being installed eg. jest and flowtype

We have worked around 1. with solutions such as Yarn resolutions and moving to
pnpm with auto-install-peers=true:

https://github.com/upleveled/eslint-config-upleveled/issues/110

Since create-react-app is going away, eslint-config-react-app will likely not be
maintained any longer (it has not received any updates for 2 years).

This commit absorbs the contents of eslint-config-react-app into our own config,
where we will maintain it from now on.